src/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-ckeditor==6.7.2
django-js-asset==3.0.1
djangorestframework==3.15.2
docopt==0.6.2
lxml==5.3.0
num2words==0.5.14
pillow==11.1.0
psycopg2==2.9.10
python-docx==1.1.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
